added base endpoint class, cleaned up types/models
