+++ src/endpoints/index.ts
@@ -1,1 +1,3 @@
 export * as Patents from './patents';
+export * from './shared-types';
+export * from './base';
