+++ src/endpoints/patents/models/investor.model.ts
@@ -1,0 +1,5 @@
+export interface Investor {
+  investor_last_name: string;
+  investor_first_name: string;
+  investor_key_id: string;
+}
