+++ src/endpoints/shared-types.ts
@@ -1,0 +1,1 @@
+export type DataPoints = string[];
